Refactor: change path from mapper package and its classes to model package
